Migrated SetupNavHostGeneratorTest.kt to DSL
